Update InputManager.cs
Changed the way mining is done. 

Input manager fires a ray from the camera every frame. (This could/should be done in another class when it's more robust)
That if the ray hits an object, pass the object to the miner if the miningKey is down.
Otherwise skip checking for mining input altogether.

Exit the mine when miningKey is no longer being pressed or when the object the mouse is currently over isn't a mineable object.