Made the Miner the primary class in mining.
Before, Mineables did more than what they should have.

Now the miner has one more variable (countdown), another event (OnProgressChange), and the routines change so that it is responsible for taking resources from a mine (as opposed to the mine 'giving' it's resources to the Miner via Miner.ReceiveResource(r). 

Where Mineable used Update to reduce mining progress while not being mined, the miner now holds a coroutine to reduce progress when leaving a mine.

The routine MineCurrent is probably the bulkiest and could be broken into other routines, however it's called every frame that the mineKey is down. It's got a few parts:

1. Checking that the mouse is over a Mineable and that it's within distance of the Miner.
2. Increasing Mining Progress.
3. Extracting resources when progress is full
3. a) Leaving the mine if empty.
4. Fire the Mining Progress event.